Update packages and code to move from Next 12 to 13
