+++ package.json
@@ -48,15 +48,15 @@
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
     "match-sorter": "^6.3.1",
-    "next": "12.3.1",
+    "next": "13",
     "next-auth": "^4.19.2",
     "next-transpile-modules": "^9.0.0",
     "notistack": "^2.0.8",
     "qs": "^6.11.2",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-csv": "^2.2.2",
     "react-device-detect": "^2.2.2",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-intl": "^6.2.0",
     "react-redux": "^8.0.4",
     "react-select": "^5.7.3",
@@ -84,8 +84,8 @@
     "@types/chance": "^1.1.3",
     "@types/lodash": "^4.14.186",
     "@types/node": "^18.7.23",
-    "@types/react": "18.0.21",
-    "@types/react-dom": "18.0.6",
+    "@types/react": "^18.2.56",
+    "@types/react-dom": "^18.2.19",
     "@types/react-table": "^7.7.12",
     "@types/react-window": "^1.8.5",
     "@types/uuid": "^8.3.4",
@@ -93,7 +93,7 @@
     "@typescript-eslint/eslint-plugin": "^5.40.0",
     "@typescript-eslint/parser": "^5.40.0",
     "eslint": "8.25.0",
-    "eslint-config-next": "12.3.1",
+    "eslint-config-next": "13",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
     "prettier": "^2.7.1",
