+++ src/pages/maintenance/500.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 
 // next
-import Image from 'next/image';
+import Image from 'next/legacy/image';
 import NextLink from 'next/link';
 
 // material-ui
